slight refactor, fix lint issues
