+++ python_modules/dagster/dagster/loggers/__init__.py
@@ -4,7 +4,7 @@ import coloredlogs
 from dagster import seven
 from dagster.config import Field
 from dagster.core.definitions.logger import logger
-from dagster.core.log_manager import coerce_valid_log_level
+from dagster.core.utils import coerce_valid_log_level
 from dagster.utils.log import default_format_string
