+++ python_modules/dagster/dagster/utils/log.py
@@ -8,7 +8,7 @@ from contextlib import contextmanager
 from dagster import check, seven
 from dagster.config import Enum, EnumValue
 from dagster.core.definitions.logger import logger
-from dagster.core.log_manager import PYTHON_LOGGING_LEVELS_MAPPING, coerce_valid_log_level
+from dagster.core.utils import PYTHON_LOGGING_LEVELS_MAPPING, coerce_valid_log_level
 
 LogLevelEnum = Enum("log_level", list(map(EnumValue, PYTHON_LOGGING_LEVELS_MAPPING.keys())))
