+++ python_modules/libraries/dagster-aws/dagster_aws/cloudwatch/loggers.py
@@ -3,7 +3,7 @@ import logging
 
 import boto3
 from dagster import Field, StringSource, check, logger, seven
-from dagster.core.log_manager import coerce_valid_log_level
+from dagster.core.utils import coerce_valid_log_level
 
 # The maximum batch size is 1,048,576 bytes, and this size is calculated as the sum of all event
 # messages in UTF-8, plus 26 bytes for each log event.
